requirements.txt
flask
flask-cors
transformers
torch
PyPDF2
python-docx
youtube-transcript-api
reportlab